refactor: 대시보드 플랫폼 타입 provider.ts로 통일
