+++ src/hooks/dashboard/usePlatformBudget.ts
@@ -1,5 +1,4 @@
 import type { TProviderType } from "@/types/dashboard/overview";
-import type { TPlatformProvider } from "@/types/dashboard/platform";
 
 import { useCoreQuery } from "@/hooks/customQuery";
 
@@ -10,12 +9,12 @@ const WARNING_THRESHOLD = 50;
 const DANGER_THRESHOLD = 75;
 
 // 단일 플랫폼 예산 소진 현황
-export function usePlatformBudget(provider: TPlatformProvider) {
+export function usePlatformBudget(provider: TProviderType) {
   const orgId = useWorkspaceStore((s) => s.selectedOrgId);
 
   return useCoreQuery(
     ["platform", "budget", orgId, provider],
-    () => getBudget(orgId!, provider as TProviderType),
+    () => getBudget(orgId!, provider),
     {
       enabled: !!orgId && !!provider,
       select: (data) => ({
